Move CodeLine
